Correctly exposed RandomUtil to Gameplay.
