+++ Source/Engine.vcxproj
@@ -696,6 +696,7 @@
     <ClCompile Include="vendors\recast\include\Recast\SampleInterfaces.cpp" />
   </ItemGroup>
   <ItemGroup>
+    <ClInclude Include="src\utils\RandomUtilDefinitions.h" />
     <ClInclude Include="resource.h" />
     <ClInclude Include="src\core\particles\modifiers\NoiseParticleModifier.h" />
     <ClInclude Include="src\core\rendering\GBuffer.h" />
